requirements.txt
Django==4.1.5
django-cors-headers == 4.3.1
djangorestframework == 3.14.0
django-environ == 0.11.2
pillow == 10.2.0